# symbol-service/api/symbol/v1/symbols_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v3.12.4
// source: symbol/v1/symbols.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationSymbolsCreateSymbol = "/symbols.v1.Symbols/CreateSymbol"
const OperationSymbolsDeleteSymbol = "/symbols.v1.Symbols/DeleteSymbol"
const OperationSymbolsGetSymbol = "/symbols.v1.Symbols/GetSymbol"
const OperationSymbolsListSymbols = "/symbols.v1.Symbols/ListSymbols"
const OperationSymbolsUpdateSymbol = "/symbols.v1.Symbols/UpdateSymbol"

type SymbolsHTTPServer interface {
	// CreateSymbol Sends a greeting
	CreateSymbol(context.Context, *CreateSymbolRequest) (*CreateSymbolResponse, error)
	DeleteSymbol(context.Context, *DeleteSymbolRequest) (*DeleteSymbolResponse, error)
	GetSymbol(context.Context, *GetSymbolRequest) (*GetSymbolResponse, error)
	ListSymbols(context.Context, *ListSymbolsRequest) (*ListSymbolsResponse, error)
	UpdateSymbol(context.Context, *UpdateSymbolRequest) (*UpdateSymbolResponse, error)
}

func RegisterSymbolsHTTPServer(s *http.Server, srv SymbolsHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/symbols", _Symbols_CreateSymbol0_HTTP_Handler(srv))
	r.GET("/v1/symbols/{id}", _Symbols_GetSymbol0_HTTP_Handler(srv))
	r.PUT("/v1/symbols/{id}", _Symbols_UpdateSymbol0_HTTP_Handler(srv))
	r.DELETE("/v1/symbols/{id}", _Symbols_DeleteSymbol0_HTTP_Handler(srv))
	r.GET("/v1/projects/{project_id}/symbols", _Symbols_ListSymbols0_HTTP_Handler(srv))
}

func _Symbols_CreateSymbol0_HTTP_Handler(srv SymbolsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateSymbolRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSymbolsCreateSymbol)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateSymbol(ctx, req.(*CreateSymbolRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateSymbolResponse)
		return ctx.Result(200, reply)
	}
}

func _Symbols_GetSymbol0_HTTP_Handler(srv SymbolsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetSymbolRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSymbolsGetSymbol)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetSymbol(ctx, req.(*GetSymbolRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetSymbolResponse)
		return ctx.Result(200, reply)
	}
}

func _Symbols_UpdateSymbol0_HTTP_Handler(srv SymbolsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateSymbolRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSymbolsUpdateSymbol)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateSymbol(ctx, req.(*UpdateSymbolRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateSymbolResponse)
		return ctx.Result(200, reply)
	}
}

func _Symbols_DeleteSymbol0_HTTP_Handler(srv SymbolsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteSymbolRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSymbolsDeleteSymbol)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteSymbol(ctx, req.(*DeleteSymbolRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteSymbolResponse)
		return ctx.Result(200, reply)
	}
}

func _Symbols_ListSymbols0_HTTP_Handler(srv SymbolsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListSymbolsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSymbolsListSymbols)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListSymbols(ctx, req.(*ListSymbolsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListSymbolsResponse)
		return ctx.Result(200, reply)
	}
}

type SymbolsHTTPClient interface {
	// CreateSymbol Sends a greeting
	CreateSymbol(ctx context.Context, req *CreateSymbolRequest, opts ...http.CallOption) (rsp *CreateSymbolResponse, err error)
	DeleteSymbol(ctx context.Context, req *DeleteSymbolRequest, opts ...http.CallOption) (rsp *DeleteSymbolResponse, err error)
	GetSymbol(ctx context.Context, req *GetSymbolRequest, opts ...http.CallOption) (rsp *GetSymbolResponse, err error)
	ListSymbols(ctx context.Context, req *ListSymbolsRequest, opts ...http.CallOption) (rsp *ListSymbolsResponse, err error)
	UpdateSymbol(ctx context.Context, req *UpdateSymbolRequest, opts ...http.CallOption) (rsp *UpdateSymbolResponse, err error)
}

type SymbolsHTTPClientImpl struct {
	cc *http.Client
}

func NewSymbolsHTTPClient(client *http.Client) SymbolsHTTPClient {
	return &SymbolsHTTPClientImpl{client}
}

// CreateSymbol Sends a greeting
func (c *SymbolsHTTPClientImpl) CreateSymbol(ctx context.Context, in *CreateSymbolRequest, opts ...http.CallOption) (*CreateSymbolResponse, error) {
	var out CreateSymbolResponse
	pattern := "/v1/symbols"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSymbolsCreateSymbol))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *SymbolsHTTPClientImpl) DeleteSymbol(ctx context.Context, in *DeleteSymbolRequest, opts ...http.CallOption) (*DeleteSymbolResponse, error) {
	var out DeleteSymbolResponse
	pattern := "/v1/symbols/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSymbolsDeleteSymbol))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *SymbolsHTTPClientImpl) GetSymbol(ctx context.Context, in *GetSymbolRequest, opts ...http.CallOption) (*GetSymbolResponse, error) {
	var out GetSymbolResponse
	pattern := "/v1/symbols/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSymbolsGetSymbol))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *SymbolsHTTPClientImpl) ListSymbols(ctx context.Context, in *ListSymbolsRequest, opts ...http.CallOption) (*ListSymbolsResponse, error) {
	var out ListSymbolsResponse
	pattern := "/v1/projects/{project_id}/symbols"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSymbolsListSymbols))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *SymbolsHTTPClientImpl) UpdateSymbol(ctx context.Context, in *UpdateSymbolRequest, opts ...http.CallOption) (*UpdateSymbolResponse, error) {
	var out UpdateSymbolResponse
	pattern := "/v1/symbols/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSymbolsUpdateSymbol))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}