# symbol-service/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Symbols API
    description: The greeting service definition.
    version: 0.0.1
paths:
    /v1/projects/{projectId}/symbols:
        get:
            tags:
                - Symbols
            operationId: Symbols_ListSymbols
            parameters:
                - name: projectId
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageToken
                  in: query
                  schema:
                    type: string
                - name: orderBy
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: sortDirection
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/symbols.v1.ListSymbolsResponse'
    /v1/symbols:
        post:
            tags:
                - Symbols
            description: Sends a greeting
            operationId: Symbols_CreateSymbol
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/symbols.v1.CreateSymbolRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/symbols.v1.CreateSymbolResponse'
    /v1/symbols/{id}:
        get:
            tags:
                - Symbols
            operationId: Symbols_GetSymbol
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/symbols.v1.GetSymbolResponse'
        put:
            tags:
                - Symbols
            operationId: Symbols_UpdateSymbol
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/symbols.v1.UpdateSymbolRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/symbols.v1.UpdateSymbolResponse'
        delete:
            tags:
                - Symbols
            operationId: Symbols_DeleteSymbol
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/symbols.v1.DeleteSymbolResponse'
components:
    schemas:
        symbols.v1.CreateSymbolRequest:
            type: object
            properties:
                projectId:
                    type: integer
                    format: int32
                uid:
                    type: string
                label:
                    type: string
                className:
                    type: string
                componentTarget:
                    type: string
                version:
                    type: string
                data:
                    type: string
                    format: bytes
            description: CREATE
        symbols.v1.CreateSymbolResponse:
            type: object
            properties:
                symbol:
                    $ref: '#/components/schemas/symbols.v1.Symbol'
        symbols.v1.DeleteSymbolResponse:
            type: object
            properties:
                success:
                    type: boolean
        symbols.v1.GetSymbolResponse:
            type: object
            properties:
                symbol:
                    $ref: '#/components/schemas/symbols.v1.Symbol'
        symbols.v1.ListSymbolsResponse:
            type: object
            properties:
                symbols:
                    type: array
                    items:
                        $ref: '#/components/schemas/symbols.v1.Symbol'
                pageToken:
                    type: string
                nextPage:
                    type: integer
                    format: int32
        symbols.v1.Symbol:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                projectId:
                    type: integer
                    format: int32
                uid:
                    type: string
                label:
                    type: string
                className:
                    type: string
                componentTarget:
                    type: string
                version:
                    type: string
                data:
                    type: string
                    format: bytes
        symbols.v1.UpdateSymbolRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                projectId:
                    type: integer
                    format: int32
                uid:
                    type: string
                label:
                    type: string
                className:
                    type: string
                componentTarget:
                    type: string
                version:
                    type: string
                data:
                    type: string
                    format: bytes
            description: UPDATE
        symbols.v1.UpdateSymbolResponse:
            type: object
            properties:
                symbol:
                    $ref: '#/components/schemas/symbols.v1.Symbol'
tags:
    - name: Symbols